[Rename] : 회원가입 페이지 폴더명 변경
- signUp -> register

resolved #9
