+++ src/pages/register/register.css
