--- src/pages/signUp/signUp.html
@@ -1,21 +1,0 @@
-<!doctype html>
-<html lang="ko-KR">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <link rel="stylesheet" href="/src/style/reset.css" />
-    <link rel="stylesheet" href="/src/style/variable.css" />
-    <link rel="stylesheet" href="/src/pages/signUp/signUp.css" />
-    <script type="module" src="/src/main.js"></script>
-    <script type="module" src="/src/pages/signUp/signUp.js"></script>
-  </head>
-  <body>
-    <h1 class="sr-only">회원가입</h1>
-    <div class="back-container">
-      <back-component></back-component>
-    </div>
-
-    <sign-up></sign-up>
-  </body>
-</html>
